Generate classes more pythonically
It's fairly standard in Python to have the __init__ method initialise the
object from given data, rather than using a separate class method as an
initialiser.  This work moves the 'dictToObject' code into the class
initialiser.

We also do a couple of simplifications:

* It's preferable not to use 'dict' as a variable name, even if it is
  actually a dictionary variable and even if Python understands that
  you're using a keyword as a variable.  So we rename this to 'config'.
* Using list comprehensions allows Python to pre-allocate the list to a
  given size, rather than appending items having to extend the list each
  iteration.
* The iterating variable doesn't have to be uniquely named, nor do we
  need a separate variable to take the data from the config that's
  then going to be iterated through.

I may be missing some reason why we used the 'dictToObject' class method
instead of standard initialisation, so please let me know if that makes
more sense.

Signed-off-by: Paul Wayper <paulway@redhat.com>